Rename config.py to models.py and remove common.py
- Rename config.py → models.py to better reflect its contents (Pydantic
  data models, not just configuration)
- Move fail_verdicts into Verdict.errors classmethod, eliminating
  common.py (which existed solely for that one function)
- Rename test_config.py → test_models.py to match
